requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
executing==2.2.0
icecream==2.1.4
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.50
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
